--- a/odoo/addons/base/tests/__init__.py
+++ b/odoo/addons/base/tests/__init__.py
@@ -20,6 +20,7 @@ from . import test_ir_actions
 from . import test_ir_attachment
 from . import test_ir_cron
 from . import test_ir_filters
+from . import test_ir_http
 from . import test_ir_mail_server
 from . import test_ir_model
 from . import test_ir_sequence