The current target, 16.0 is a little arbitrary, to discuss.

The routing map is an important part of the cold loading of a page. This can be improved multiple way.

## Possible improvements:
### Improve the totally cold routing map
Most of the time spent is in the `werkzeug.routing.Rule.add` method representing 70~80% of the total routing map generation time.
This will need some low level thinking and optimisations
Another part is the `odoo.http._generate_routing_rules`, not so massive but still visible, mainly the `getmembers`, `is_method_a_route`, `build_controllers`, ... Hard to say what is possible to do but a possible improvement would be to generate a registry of routes inside the route decorator.

### Improve muti webiste routing map
Between website, the main difference is the rewrites rules:
`rewrites = dict([(x.url_from, x) for x in request.env['website.rewrite'].sudo().search(domain)])`

A simple assumption is that, if this list is empty, we could use the `None` routing map. 
Even if this list is not empty, we could alter the `None` routing map to remove some rules, add some redirect
A ModelConverter is also added if website is installed.

### Improve multidb routing map
The mro dynamic composition would make this difficult but investigating a per module routing map and combining them
based on the installed module could be an important improvement for the saas infrastructure. 

## What is actually done in this pr
Even if there are more possible improvements, werkzeug 2.2 will change how the routing map is managed, going from regexes to a decision tree. The change proposed in this pr change should remain compatible in theory or at least not to difficult to adapt.

### Lazy _compile_builder: 
~600 ms -> ~200ms. 
Instead of generating the build method for ~1100 routes, will only create the method when accessing this page and matching the rule, or when accessing the sitemap or searching a page (only a small amount of routes are actually build in this scenario)

### Increase re _MAX_SIZE
One important part of the routing map is compiling the regex for routes. They actually are the same for all databases, all website, .... and it take around 100~150 ms. Last test increasing the _MAX_SIZE from 512 to 2200 allowed to speedup the 
second routing map from ~200 to ~70 ms. This improvement is also cross db, but 2200 is maybe to low to benefit from that
since the re._MAX_SIZE is used by all regexes. The proposed size here is 4096.
This imp should be visible either for a db with mutiple website, or for a worker serving multiple db, or even if the routing map was invalidated and regenared.

Note that this cache could be conceptually dedicated to routes and populated exclusively from the @route decorator, avoiding to add database specific rules to this cache, but this is a little hard to combine.