@rdeodoo 2023-05-03
Note that I just read the commit message, this might be obvious or off-topic.
I am just curious about:
> Instead of generating the build method for ~1100 routes, will only create the method when accessing this page and matching the rule, or when accessing the sitemap or searching a page (only a small amount of routes are actually build in this scenario)

1. In other words, does that mean that we are now improving the cold loading perfs at the cost of perfs in hot state when accessing a page? (So, instead of encapsulated only once until next invalidation or server restart, it will be spread when accessing a page, maybe hours/days later?)
Or maybe I understood it wrong and it was actually done in both cases anyway and this is just removing the useless usage during routing map generation?
2. If yes:
2.1. Is it only once per same page access (so could even be hour or day later when a page is first reached), or every time a same page is accessed?
2.2. Do you have metric about the perf loss on those page access?