@Xavier-Do 2023-05-04
> 1. In other words, does that mean that we are now improving the cold loading perfs at the cost of perfs in hot state when accessing a page? (So, instead of encapsulated only once until next invalidation or server restart, it will be spread when accessing a page, maybe hours/days later?)

There will still be an hot state, but the cache will populate only on demand for what is needed than populating the whole cache at once on the first page loaded.

>    Or maybe I understood it wrong and it was actually done in both cases anyway and this is just removing the useless usage during routing map generation?

With this solution, some useless build method wont be generated anymore (technical routes, backend route) Only frontend route will be generated in fine and only when requested.

> 2. If yes:
>    2.1. Is it only once per same page access (so could even be hour or day later when a page is first reached), or every time a same page is accessed?

Not sure I understand your question but if you access /event/<event> for the first time, the build method will be added to the routing map and never computed again until the routing map is invalidated. If the page /contact is accessed for the first time three days latter, this build method will be computed at this time and also added to the routing map. But I don't really expect a routing map to live for such a long time.

2.2. Do you have metric about the perf loss on those page access?
It is hard to say but this is would be round 400ms/1100 per page on first access on average, less than 0.4 ms, only on the first access of a few page, but still gaining 400 ms on the first accessed page of a worker when generating the routing map. But for most pages I assume that it would be even faster than that, and only take time for more complex routes.