@Julien00859 2023-05-04
Looks like `/event/<model("event.event"):event>/booth` was build twice:

```
WARNING odoo.addons.base.models.ir_http: building /event/<model("event.event"):event>/booth in 0.00034 
INFO werkzeug: "GET /event/design-fair-los-angeles-1/booth HTTP/1.1" 200 - 78 0.027 0.093
```

```
WARNING odoo.addons.base.models.ir_http: building /event/<model("event.event"):event>/booth in 0.00039 
[...]
INFO werkzeug: "GET /sitemap.xml HTTP/1.1" 200 - 256 0.150 0.428
```

Is it normal, is the sitemap different?